1. Material 수정 2. Obj Importer 작업 중
나머지 작업은, 다른 밑 작업을 마저하고 작업할 예정

re #17

+++ SOCEngine/Platform/Win32/TestScene.cpp
@@ -69,7 +69,7 @@ void TestScene::OnInitialize()
 				Math::Vector3( -1.0f, 1.0f, 1.0f )
 			};
 
-			static WORD indices[] =
+			static unsigned int indices[] =
 			{
 				3,1,0,
 				2,1,3,
